[git] 페이지 분리
